split protractor configuration file
