+++ package.json
@@ -144,7 +144,6 @@
     "husky": "^1.2.0",
     "jasmine-ajax": "3.2.0",
     "jasmine-core": "~2.8.0",
-    "jasmine-fail-fast": "^2.0.0",
     "jasmine-reporters": "^2.3.2",
     "jasmine-spec-reporter": "~4.2.1",
     "jasmine2-protractor-utils": "1.1.3",
@@ -163,6 +162,7 @@
     "license-checker": "^25.0.1",
     "lint-staged": "^8.1.7",
     "lite-server": "^2.4.0",
+    "lodash.wrap": "^4.1.1",
     "markdownlint-cli": "^0.16.0",
     "mdast-util-compact": "^1.0.1",
     "mdast-util-heading-range": "^2.1.0",
